% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_psi_cpp}
\alias{calc_psi_cpp}
\title{Calculate Aggregated Projection Matrix (Core for APM)}
\usage{
calc_psi_cpp(K_list)
}
\arguments{
\item{K_list}{A list of matrices (T x r) representing factor spaces for each group.}
}
\value{
A matrix (T x T) representing the average projection matrix.
}
\description{
Calculate Aggregated Projection Matrix (Core for APM)
}
