% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpe.R
\name{cpe}
\alias{cpe}
\title{Circular Projection Estimator (CPE) for Comparison}
\usage{
cpe(Y_list, r0 = 2, r_max = 8, max_iter = 100)
}
\arguments{
\item{Y_list}{A list of matrices (T x N_m).}

\item{r0}{The number of global factors to estimate.}

\item{r_max}{The dimension of the subspace to project onto in each group.}

\item{max_iter}{Maximum iterations.}
}
\value{
A list containing the estimated global factors `G_hat`.
}
\description{
Implements the iterative projection method (Chen, 2022) to serve as a benchmark.
}
