% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler_cpp}
\alias{gibbs_sampler_cpp}
\title{A Gibbs Sampler using Rcpp}
\usage{
gibbs_sampler_cpp(n_iter, y, start_theta)
}
\arguments{
\item{n_iter}{Number of iterations}

\item{y}{Input vector}

\item{start_theta}{Initial value}
}
\value{
A vector of samples
}
\description{
A Gibbs Sampler using Rcpp
}
