% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA25204152-package.R
\docType{package}
\name{SA25204152-package}
\alias{SA25204152}
\alias{SA25204152-package}
\title{SA25204152: Final Project for Statistical Computing}
\description{
Homework and research functions for the course.
}
\keyword{internal}
