% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_group_data}
\alias{sim_group_data}
\title{Generate Simulation Data for Group Factor Model}
\usage{
sim_group_data(
  M = 3,
  N = 50,
  T_obs = 100,
  r0 = 2,
  r_local = 2,
  phi_G = 0.5,
  phi_F = 0.5,
  phi_e = 0.5
)
}
\arguments{
\item{M}{Number of groups.}

\item{N}{Number of individuals per group (can be a vector or scalar).}

\item{T_obs}{Number of time observations.}

\item{r0}{Number of global factors.}

\item{r_local}{Number of local factors per group (scalar).}

\item{phi_G}{AR(1) coefficient for global factors.}

\item{phi_F}{AR(1) coefficient for local factors.}

\item{phi_e}{AR(1) coefficient for idiosyncratic errors.}
}
\value{
A list containing `Y_list` (data) and true factors/loadings.
}
\description{
Generates data based on the settings in Li et al. (2024), Slide 35.
}
